requirements.txt
discord.py[voice]
openai
flask
asyncio
youtube_dl
ffmpeg-python
beautifulsoup4
requests